+++ ext/tiny_tds/client.c
@@ -137,9 +137,8 @@ static void dbcancel_ubf(DBPROCESS *client)
   userdata->dbcancel_sent = 1;
 }
 
-static void nogvl_setup(DBPROCESS *client)
+static void nogvl_setup(tinytds_client_userdata *userdata)
 {
-  GET_CLIENT_USERDATA(client);
   userdata->nonblocking = 1;
   userdata->nonblocking_errors_length = 0;
   userdata->nonblocking_errors = malloc(ERRORS_STACK_INIT_SIZE * sizeof(tinytds_errordata));
@@ -183,8 +182,9 @@ static void nogvl_cleanup(DBPROCESS *client)
 
 static RETCODE nogvl_dbnextrow(DBPROCESS * client)
 {
+  GET_CLIENT_USERDATA(client);
   int retcode = FAIL;
-  nogvl_setup(client);
+  nogvl_setup(userdata);
   retcode = NOGVL_DBCALL(dbnextrow, client);
   nogvl_cleanup(client);
   return retcode;
@@ -192,8 +192,9 @@ static RETCODE nogvl_dbnextrow(DBPROCESS * client)
 
 static RETCODE nogvl_dbresults(DBPROCESS *client)
 {
+  GET_CLIENT_USERDATA(client);
   int retcode = FAIL;
-  nogvl_setup(client);
+  nogvl_setup(userdata);
   retcode = NOGVL_DBCALL(dbresults, client);
   nogvl_cleanup(client);
   return retcode;
@@ -201,8 +202,9 @@ static RETCODE nogvl_dbresults(DBPROCESS *client)
 
 static RETCODE nogvl_dbsqlexec(DBPROCESS *client)
 {
+  GET_CLIENT_USERDATA(client);
   int retcode = FAIL;
-  nogvl_setup(client);
+  nogvl_setup(userdata);
   retcode = NOGVL_DBCALL(dbsqlexec, client);
   nogvl_cleanup(client);
   return retcode;
@@ -212,7 +214,7 @@ static RETCODE nogvl_dbsqlok(DBPROCESS *client)
 {
   int retcode = FAIL;
   GET_CLIENT_USERDATA(client);
-  nogvl_setup(client);
+  nogvl_setup(userdata);
   retcode = NOGVL_DBCALL(dbsqlok, client);
   nogvl_cleanup(client);
   userdata->dbsqlok_sent = 1;
@@ -292,6 +294,17 @@ static void push_userdata_error(tinytds_client_userdata *userdata, tinytds_error
   userdata->nonblocking_errors_length++;
 }
 
+typedef struct {
+  DBPROCESS *dbproc;
+  tinytds_errordata *error_data;
+} raise_error_args_t;
+
+static void* raise_error_with_gvl(void *ptr)
+{
+  raise_error_args_t *args = (raise_error_args_t *)ptr;
+  rb_tinytds_raise_error(args->dbproc, *(args->error_data));
+}
+
 int tinytds_err_handler(DBPROCESS *dbproc, int severity, int dberr, int oserr, char *dberrstr, char *oserrstr)
 {
   static const char *source = "error";
@@ -369,6 +382,17 @@ int tinytds_err_handler(DBPROCESS *dbproc, int severity, int dberr, int oserr, c
     }
 
     push_userdata_error(userdata, error_data);
+  } else if (!userdata) {
+    // userdata can only really be NULL if something goes wrong during dbopen
+    // and dbopen is also executed as non-blocking
+    // so we have to re-acquire the GVL in order to directly raise a Ruby error
+    // in newer Ruby version we seem to have a dedicated method vailable to simplify the statement here
+    // https://docs.ruby-lang.org/capi/en/master/d6/dfb/include_2ruby_2thread_8h.html#a2293d6040c352991d160113a62fe5be3
+    raise_error_args_t args;
+    args.dbproc = dbproc;
+    args.error_data = &error_data;
+
+    rb_thread_call_with_gvl(raise_error_with_gvl, &args);
   } else {
     rb_tinytds_raise_error(dbproc, error_data);
   }
@@ -936,6 +960,16 @@ static void *dbuse_without_gvl(void *ptr)
   return NULL;
 }
 
+struct dbopen_args {
+  LOGINREC *login;
+  const char *dataserver;
+};
+
+static void * dbopen_without_gvl(void *ptr)
+{
+  struct dbopen_args *args = (struct dbopen_args *)ptr;
+  return dbopen(args->login, args->dataserver);
+}
 
 static VALUE rb_tinytds_connect(VALUE self)
 {
@@ -1005,9 +1039,29 @@ static VALUE rb_tinytds_connect(VALUE self)
     DBSETLUTF16(cwrap->login, 0);
   }
 
-  cwrap->client = dbopen(cwrap->login, StringValueCStr(dataserver));
+  struct dbopen_args open_args;
+
+  open_args.login = cwrap->login;
+
+  open_args.dataserver = StringValueCStr(dataserver);
+
+  nogvl_setup(cwrap->userdata);
+
+  cwrap->client = (DBPROCESS *)rb_thread_call_without_gvl(
+                    dbopen_without_gvl,
+                    &open_args,
+                    RUBY_UBF_IO,
+                    0
+                  );
 
   if (cwrap->client) {
+    // nogvl_cleanup will retrieve the user data via the client
+    // as the client will be used to find out details about an error
+    // in order for the client to have a pointer to our userdata for information
+    // dbsetuserdata has to be called first
+    dbsetuserdata(cwrap->client, (BYTE*)cwrap->userdata);
+    nogvl_cleanup(cwrap->client);
+
     if (dbtds(cwrap->client) < 11) {
       rb_raise(cTinyTdsError, "connecting with a TDS version older than 7.3!");
     }
@@ -1028,7 +1082,6 @@ static VALUE rb_tinytds_connect(VALUE self)
       }
     }
 
-    dbsetuserdata(cwrap->client, (BYTE*)cwrap->userdata);
     dbsetinterrupt(cwrap->client, check_interrupt, handle_interrupt);
     cwrap->userdata->closed = 0;
 
@@ -1039,18 +1092,20 @@ static VALUE rb_tinytds_connect(VALUE self)
 
       // in case of any errors, the tinytds_err_handler will be called
       // so we do not have to check the return code here
-      nogvl_setup(cwrap->client);
+      nogvl_setup(cwrap->userdata);
       rb_thread_call_without_gvl(
         dbuse_without_gvl,
         &use_args,
-        NULL,
-        NULL
+        RUBY_UBF_IO,
+        0
       );
       nogvl_cleanup(cwrap->client);
     }
 
     cwrap->encoding = rb_enc_find(StringValueCStr(charset));
     cwrap->identity_insert_sql = "SELECT CAST(SCOPE_IDENTITY() AS bigint) AS Ident";
+  } else {
+    rb_raise(cTinyTdsError, "Unable to connect to SQL Server!");
   }
 
   return self;